% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My confidence interval function}
\usage{
myci(y, alpha = 0.05)
}
\arguments{
\item{y}{the sample data}

\item{alpha}{the alpha value}
}
\value{
a list containing the confidence interval
}
\description{
Function to form a confidence interval
}
\details{
This is a helper function used in Lab 11 of Math4753 that gives a nice and simple way to calculate a confidence interval.
}
\examples{
set.seed(23);x=rnorm(30,mean=10,sd=12)
myci(x)
}
