% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{size}

\item{iter}{iterations}

\item{a}{5}

\item{b}{10}
}
\value{
a plot
}
\description{
myclt
}
\examples{
\dontrun{w=myclt(n=50,iter=10000,a=5,b=10)}
}
