% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param)
}
\arguments{
\item{lfun}{log function}

\item{x}{x variable}

\item{param}{parameter}
}
\value{
a list
}
\description{
mymaxlik
}
\examples{
\dontrun{mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin)}
}
