% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My normal curve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{the value to calculate prob with}
}
\value{
the probability
}
\description{
Plots a normal curve and the probability of X<=a
}
\details{
This is a normal curve plotting function
}
\examples{
myncurve(10,5,6)
}
