% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypie.R
\name{mypie}
\alias{mypie}
\title{My pie chart function}
\usage{
mypie(slices, lbls, mn)
}
\arguments{
\item{slices}{the slices for the pie chart}

\item{lbls}{the labels for the slices}

\item{mn}{the main title of the pie chart}
}
\value{
a pie chart
}
\description{
Creates a pie chart with percentages
}
\details{
This is a helper function for creating a nice pie chart with percentages for each slice
}
\examples{
slices <- c(20,18)
lbls <- c("slice 1", "slice 2")
mypie(slices, lbls, "mypie")
}
