% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My pvalue function}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the t value}

\item{xmax}{the maximum x value}

\item{n}{the size of data}

\item{alpha}{the alpha value}
}
\value{
list containing quantiles and pvalue
}
\description{
This function displays p-value areas
}
\details{
This is a helper function used in Lab 12 that calculates and displays p value areas
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
tcalc=(xbar-24)/(sd(x1)/sqrt(length(x1)))
mypvalue(tcalc,n=30,alpha=0.05)
}
