% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{My sample function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{number of iterations}

\item{time}{seconds to timeout between plots}
}
\value{
the sample factor
}
\description{
Creates and plots a sample of x
}
\details{
This is a test function for taking samples
}
\examples{
mysample(n=1000,iter=30,time=1)
}
