% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{My pareto function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{a vector}
}
\value{
pareto diagram
}
\description{
Creates a pareto diagram
}
\details{
This is a helper function for creating pareto diagrams
}
\examples{
x = 1:20
pareto(x)
}
