% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.R
\name{rsq}
\alias{rsq}
\title{My r squared function}
\usage{
rsq(xk, data)
}
\arguments{
\item{xk}{the xk value}

\item{data}{the data frame}
}
\value{
a vector of doubles
}
\description{
Creates the r squared value
}
\details{
This is a helper function for calculatiing the r squared value of a linear
regression model.
}
\examples{
xk <- 18
data <- 1:20/20
rsq(xk, data)
}
